requirements.txt
pandas==2.2.3
protobuf==5.29.5
python-dotenv==1.1.0
streamlit==1.45.1
google-adk==1.1.1
google-ai-generativelanguage==0.6.15
google-cloud-storage==2.19.0
google-cloud-bigquery==3.34.0